pyproject.toml
[project]
name = "hr-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-openai",
    "streamlit",
    "python-dotenv",
    "numpy",
    "pandas",
    "notebook<7",
    "langchain-ollama>=1.0.1",
    "langfuse>=4.0.1",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "nbclassic>=1.3.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "psycopg[binary]>=3.1.0",
    "aiosqlite>=0.20.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "openpyxl>=3.1.5",
]
